refactor(admin-stats): route formatCount through local formatters barrel for intra-module consistency
